number of files

+++ class_write_file_and_decode.py
@@ -21,29 +21,31 @@ def readQR(image):
         data = qr[0].data.decode("utf-8")
         return data
 recieve_folder = "recieved/"
+
+
 def writefile(scanned_data):
-    print("recieve:")
-    print(scanned_data)
+    previous_number = 0
     metadata_index = scanned_data.index("\n&&&&&&&&&&&&777777777777\n") + 1
     metadata_recieved = scanned_data[:metadata_index]
-    string_recieved = scanned_data[metadata_index + 28:len(scanned_data)]
+    string_recieved = scanned_data[metadata_index + 25:len(scanned_data)]
     metadata_json = json.loads(metadata_recieved, object_hook=customStudentDecoder)
     number_of_file = int(metadata_json.p)
     number_all_of_files = int(metadata_json.a)
-
     original_filename = str(metadata_json.f).replace("\\", "/").replace("¥¥", "/")
     filename = recieve_folder + ('part%04d' % number_of_file)
-    shared_utilites.write_file(filename, string_recieved)
+    if previous_number != number_of_file:
+        previous_number = number_of_file
+        shared_utilites.write_file(filename, string_recieved)
 
-    if number_of_file == number_all_of_files:
-        # end of transportation
-        final_file_location = "recieved_files\\" + original_filename
-        if os.path.isfile(final_file_location):
-            os.remove(final_file_location)
-        class_join.join.join(recieve_folder, final_file_location)
-        if os.path.isdir(recieve_folder):
-            for eachFile in shared_utilites.load_splitted_files(recieve_folder):
-                os.remove(recieve_folder + "/" + eachFile)
+        if number_of_file == number_all_of_files:
+            # end of transportation
+            final_file_location = "recieved_files\\" + original_filename
+            if os.path.isfile(final_file_location):
+                os.remove(final_file_location)
+            class_join.join.join(recieve_folder, final_file_location)
+            if os.path.isdir(recieve_folder):
+                for eachFile in shared_utilites.load_splitted_files(recieve_folder):
+                    os.remove(recieve_folder + "/" + eachFile)
         # move_file
 
 
@@ -51,6 +53,7 @@ def writefile(scanned_data):
 
         # os.rename(filename,"folder/"+original_filename)
 class write_file_and_deocde:
+    previous_number = 0
     def decodeimag(img):
         scanned_data = readQR(img)
         writefile(scanned_data)
